feat: 10日再生数更新処理を別Functionに分離しスケール耐性を改善
- 新規: backend/functions/service/data_sync/services/ten_days_metrics_update.py
  - 各テーブルの10日指標更新を担当
  - 成功後 play-count-correction にPublish
- 変更: backend/functions/service/data_sync/services/video_history_sync.py
  - 履歴同期のみ実行し、update-ten-days-metrics トピックにPublishするよう変更
- 修正: backend/functions/service/data_sync/main.py
  - followers_update の import エイリアス修正
  - update_ten_days_metrics_function をエクスポート
- インフラ: backend/functions/service/data_sync/cloudbuild.yaml
  - update-ten-days-metrics 関数のデプロイ設定とトリガートピックを整合

処理チェーン: video-history-sync → update-ten-days-metrics → play-count-correction
