+++ backend/functions/service/data_sync/main.py
@@ -15,6 +15,7 @@ from services.sync_corporate_data import sync_corporate_data as sync_corporate_d
 from services.play_count_correction import correct_play_count_increase as play_count_correction_function
 from services.sync_second_third_account_type import sync_second_third_account_type as sync_second_third_account_type_function
 from services.followers_update import update_followers as update_followers_function
+from services.ten_days_metrics_update import update_ten_days_metrics as update_ten_days_metrics_function
 
 # 環境変数の読み込み
 load_dotenv()
@@ -60,3 +61,6 @@ def update_followers(event, context):
 
 def sync_second_third_account_type(event, context):
     return sync_second_third_account_type_function(event, context)
+
+def update_ten_days_plays_metrics(event, context):
+    return update_ten_days_metrics_function(event, context)
